test: cleanup QUIC tests and migrate them to mjs
