+++ test/common/index.mjs
@@ -16,6 +16,7 @@ const {
   getBufferSources,
   getTTYfd,
   hasCrypto,
+  hasQuic,
   hasSQLite,
   hasIntl,
   hasIPv6,
@@ -66,6 +67,7 @@ export {
   getPort,
   getTTYfd,
   hasCrypto,
+  hasQuic,
   hasSQLite,
   hasIntl,
   hasIPv6,
